feat: adding Hebrew translations for legacy footers
in order to prove that we are using our forked version of the translations
repository,  I wanted to add some translations that don't exist in
transifex.  I chose Hebrew because (1) I can read it, and (2)
translation strings exist elsewhere in  this repo, for other pages, so I
didn't have to re-create from scratch.